feat(interfaces): organize interfaces into domain-specific modules
