+++ src/adobe/Adobe.ts
@@ -1,6 +1,6 @@
-import { Mail } from '../mail/Mail'
-import { Account } from './Account'
+import {Mail} from '../mail/Mail'
+import {Account} from './Account'
 
 export interface Adobe {
-    newAccount(mail: Mail): Promise<Account>
+  account(mail: Mail): Promise<Account>
 }
