+++ src/mail/Mail.ts
@@ -1,5 +1,7 @@
 export interface Mail {
-    address(): Promise<string>
-    password(): Promise<string>
-    delete(): Promise<void>
+  address(): Promise<string>
+
+  password(): Promise<string>
+
+  delete(): Promise<void>
 }
