+++ src/session/Session.ts
@@ -1,10 +1,8 @@
-import { Mail } from '../mail/Mail'
-import {Json, SessionId} from '../aliases/Aliases'
+import {Account} from '../adobe/Account'
+import {Json} from '../Aliases'
 
 export interface Session {
-    id(): Promise<SessionId>
-    mail(): Promise<Mail>
-    update(mail: Mail): Promise<void>
-    valid(): Promise<boolean>
-    asJson(): Promise<Json>
+  account(): Promise<Account>
+  asJson(): Promise<Json>
+  delete(): Promise<void>
 }
